pyproject.toml
[project]
name = "blog-content-automation"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "pdfminer-six (>=20250506,<20250507)",
    "pypdf2 (>=3.0.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
